@@ -14,26 +14,27 @@
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.16",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "avataaars": "^2.0.0",
-    "bootstrap": "^5.1.3",
-    "gatsby": "^4.5.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.1.0",
-    "react-dom": "^17.0.2",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-brands-svg-icons": "^6.4.2",
+    "@fortawesome/free-solid-svg-icons": "^6.4.2",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mui/icons-material": "^5.14.3",
+    "@mui/material": "^5.14.4",
+    "bootstrap": "^5.3.1",
+    "framer-motion": "^10.16.0",
+    "gatsby": "^5.11.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
     "react-fontawesome": "^1.7.1",
-    "react-icons": "^4.3.1",
-    "react-reveal": "^1.2.2",
-    "react-router-dom": "^6.2.1",
-    "react-scroll": "^1.8.4"
+    "react-icons": "^4.10.1",
+    "react-router-dom": "^6.14.2",
+    "react-scroll": "^1.8.9"
   },
   "devDependencies": {
-    "prettier": "^2.5.1"
+    "prettier": "^3.0.1"
   },
   "repository": {
     "type": "git",
