@@ -1,37 +1,48 @@
 import React from "react";
-import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-import {faGithub} from "@fortawesome/free-brands-svg-icons";
-import {IconButton, Typography, Accordion, AccordionDetails, AccordionSummary} from "@material-ui/core"
-import {ExpandMore} from "@material-ui/icons";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faGithub } from "@fortawesome/free-brands-svg-icons";
+import { Card, CardContent, CardActions, IconButton, Typography } from "@mui/material"
+import Reveal from './animation/reveal';
+import Grid from '@mui/material/Unstable_Grid2'; // Grid version 2
 
-const AccordionItem = (props) =>
-        <Accordion style={{backgroundColor: "#f7fcff", width: '85%'}}>
-            <AccordionSummary
-                expandIcon={<ExpandMore />}
-            >
-                <Typography>{props.title}</Typography>
-            </AccordionSummary>
-            <AccordionDetails align="left" style={{display: "flex",
-                flexDirection: "column"}}>
-                <Typography variant={"subtitle2"}>
-                    {props.language}
-                    <IconButton href={props.link}>
-                        <FontAwesomeIcon icon={faGithub}/>
-                    </IconButton>
-                </Typography>
-                <Typography>
-                    {props.body}
-                </Typography>
-            </AccordionDetails>
-        </Accordion>
+const CardItem = (props) =>
+    <Card className="card">
+        <CardContent>
+            <Typography className="card-title" variant="h5">{props.title}</Typography>
+            <Typography sx={{ mb: 1.5 }} color="text.secondary">
+                {props.language}
+            </Typography>
+            <Typography variant="body2">
+                {props.body}
+            </Typography>
+        </CardContent>
+        <CardActions><IconButton href={props.link}>
+            <FontAwesomeIcon icon={faGithub} />
+        </IconButton></CardActions>
+    </Card>
 
-const ProjectsComponent = () =>
+const Projects = () =>
     <div id="projects">
-            <AccordionItem title={"personal website"} language={"GatsbyJS, MaterialUI, HTML/CSS, JS"} body={"You're looking at it! For this project, my goals were to create a simple, one-page website that is both\n" +
-            "desktop and mobile-friendly (WIP)."} link={"https://github.com/shalupova-a/angelina.shalupova"}/>
-            <AccordionItem title={"corvus web app"} language={"HTML/CSS, JS, ReactJS, Redux, Java, Bootstrap"} body={"This project started as a group effort for CS4550 - Web Development, and I have been working independently to improve it.\n" +
-            "Combining my love for astrology and music, this website recommends songs based on one's astrological sun sign!"} link={"https://github.com/shalupova-a/corvus"}/>
-            <AccordionItem title={"snarl game"} language={"Java"} body={"This project was a group effort for CS4500 - Software Development. This is a server-client 2D dungeon crawler game!"} link={"https://github.com/shalupova-a/Snarl"}/>
+        <Reveal>
+            <Grid container sx={{ marginTop: '5%' }}>
+                <Grid align="left" item xs={6}>
+                    <p className="title">projects</p>
+                </Grid>
+                <Grid container xs={6} spacing={3}>
+                    <Grid item xs={6}>
+                        <CardItem title={"Personal Website"} language={"GatsbyJS, MaterialUI, React, HTML/CSS, JS"} body={"You're looking at it! For this project, my goals were to create a simple, one-page website that is both\n" +
+                            "desktop and mobile-friendly (WIP)."} link={"https://github.com/shalupova-a/angelina.shalupova"} />
+                    </Grid>
+                    <Grid item xs={6}>
+                        <CardItem title={"Corvus"} language={"HTML/CSS, JS, ReactJS, Redux, Java, Bootstrap"} body={"This project started as a group effort for CS4550 - Web Development.\n" +
+                            "This website recommends songs based on one's astrological sun sign!"} link={"https://github.com/shalupova-a/corvus"} />
+                    </Grid>
+                    <Grid item xs={6}>
+                        <CardItem title={"Snarl"} language={"Java"} body={"This project was a group effort for CS4500 - Software Development. This is a server-client 2D dungeon crawler game!"} link={"https://github.com/shalupova-a/Snarl"} />
+                    </Grid>
+                </Grid>
+            </Grid>
+        </Reveal>
     </div>
 
-export default ProjectsComponent
+export default Projects
