@@ -1,14 +1,15 @@
 import React from "react";
-import {Grid} from "@material-ui/core"
-import {FaGithub, FaLinkedin, FaEnvelope} from "react-icons/fa"
+import {Grid} from "@mui/material"
+import {FaGithub, FaLinkedin, FaEnvelope, FaFileAlt} from "react-icons/fa"
 
-const SocialsComponent = () =>
+const Socials = () =>
     <div id="socials">
         <Grid
             container
             item
             justify="space-evenly"
-            alignItems="flex-start">
+            align="right"
+            spacing={4}>
             <Grid item>
                 <a href="https://github.com/shalupova-a">
                     <FaGithub className="social-icon" size={25}/>
@@ -20,7 +21,12 @@ const SocialsComponent = () =>
                 </a>
             </Grid>
             <Grid item>
-                <a href="mailto:shalupova.a@northeastern.edu">
+                <a href="https://ashalupova-resume.tiiny.site">
+                    <FaFileAlt size={25} />
+                </a>
+            </Grid>
+            <Grid item>
+                <a href="mailto:a.shalupova@gmail.com">
                     <FaEnvelope className="social-icon" size={25}/>
                 </a>
             </Grid>
@@ -29,4 +35,4 @@ const SocialsComponent = () =>
         <br/>
     </div>
 
-export default SocialsComponent
+export default Socials
